requirements.txt
numpy>=1.19.2
torch>=1.7.0
PyYAML>=5.3.1
rdkit-pypi>=2020.09.1
scikit-learn>=0.23.2
